Fixed warnings reported by gcc

+++ source/memory/OamManager.cpp
@@ -88,7 +88,8 @@ OAMManager::initSprite(const unsigned short pallette[], int palLen, const unsign
     /* Keep track of the available tiles */
 //    assert(current_oam_id_palette < SPRITE_COUNT && current_oam_id_tiles < SPRITE_COUNT);
     if (!(current_oam_id_palette < SPRITE_COUNT && current_oam_id_tiles < SPRITE_COUNT)) {
-        printf("\n\n\nTOO MUCH creatures/PALETTES! %i %i", current_oam_id_palette, current_oam_id_tiles);
+        printf("\n\n\nTOO MUCH creatures/PALETTES! %lu %lu",
+               static_cast<unsigned long>(current_oam_id_palette), static_cast<unsigned long>(current_oam_id_tiles));
         for (int a = 0; a < 120; a++) swiWaitForVBlank();
         return nullptr;
     }
